Refactor of the interpreter and the internals for better testing. Fixes #46 and #37.
