+++ core/src/test/scala/com/github/gvolpe/fs2rabbit/interpreter/AMQPClientInMemory.scala
@@ -18,6 +18,7 @@ package com.github.gvolpe.fs2rabbit.interpreter
 
 import cats.effect.IO
 import com.github.gvolpe.fs2rabbit.algebra.AMQPClient
+import com.github.gvolpe.fs2rabbit.config.Fs2RabbitConfig
 import com.github.gvolpe.fs2rabbit.model
 import com.github.gvolpe.fs2rabbit.model._
 import com.github.gvolpe.fs2rabbit.model.ExchangeType.ExchangeType
@@ -25,32 +26,46 @@ import com.rabbitmq.client.Channel
 import fs2.Stream
 import fs2.async.mutable
 
+import scala.collection.mutable.{Set => MutableSet}
+
 class AMQPClientInMemory(internalQ: mutable.Queue[IO, Either[Throwable, AmqpEnvelope]],
-                         ackerQ: mutable.Queue[IO, AckResult]) extends AMQPClient[Stream[IO, ?]] {
+                         ackerQ: mutable.Queue[IO, AckResult],
+                         config: IO[Fs2RabbitConfig]) extends AMQPClient[Stream[IO, ?]] {
+
+  private val queues: MutableSet[QueueName] = MutableSet.empty[QueueName]
 
   override def basicAck(channel: Channel, tag: model.DeliveryTag, multiple: Boolean): Stream[IO, Unit] = {
     Stream.eval(ackerQ.enqueue1(Ack(tag)))
   }
   override def basicNack(channel: Channel, tag: model.DeliveryTag, multiple: Boolean, requeue: Boolean): Stream[IO, Unit] = {
-    Stream.eval(ackerQ.enqueue1(NAck(tag)))
+    // Imitating the RabbitMQ behavior
+    val envelope = AmqpEnvelope(DeliveryTag(1), "requeued msg", AmqpProperties.empty)
+    for {
+      _ <- Stream.eval(ackerQ.enqueue1(NAck(tag)))
+      c <- Stream.eval(config)
+      _ <- if (c.requeueOnNack) Stream.eval(internalQ.enqueue1(Right(envelope))) else Stream.eval(IO.unit)
+    } yield ()
   }
   override def basicQos(channel: Channel, basicQos: model.BasicQos): Stream[IO, Unit] = Stream.eval(IO.unit)
 
   override def basicConsume(channel: Channel, queueName: model.QueueName, autoAck: Boolean, consumerTag: String, noLocal: Boolean, exclusive: Boolean, args: Map[String, AnyRef]): Stream[IO, String] = {
-    Stream.eval(IO("dequeue1 happens in AckerConsumerProgram.createConsumer"))
+    val ifError = Stream.raiseError[String](new java.io.IOException(s"Queue ${queueName.value} does not exist!")).covary[IO]
+    queues.find(_.value == queueName.value).fold(ifError) { _ =>
+      Stream.eval(IO("dequeue1 happens in AckerConsumerProgram.createConsumer"))
+    }
   }
   override def basicPublish(channel: Channel, exchangeName: model.ExchangeName, routingKey: model.RoutingKey, msg: model.AmqpMessage[String]): Stream[IO, Unit] = {
     val envelope = AmqpEnvelope(DeliveryTag(1), msg.payload, msg.properties)
     Stream.eval(internalQ.enqueue1(Right(envelope)))
   }
 
-  override def deleteQueue(channel: Channel, queueName: model.QueueName, ifUnused: Boolean, ifEmpty: Boolean): Stream[IO, Unit] = Stream.eval(IO.unit)
-  override def deleteQueueNoWait(channel: Channel, queueName: model.QueueName, ifUnused: Boolean, ifEmpty: Boolean): Stream[IO, Unit] = Stream.eval(IO.unit)
+  override def deleteQueue(channel: Channel, queueName: model.QueueName, ifUnused: Boolean, ifEmpty: Boolean): Stream[IO, Unit] = Stream.eval(IO(queues -= queueName)).map(_ => ())
+  override def deleteQueueNoWait(channel: Channel, queueName: model.QueueName, ifUnused: Boolean, ifEmpty: Boolean): Stream[IO, Unit] = Stream.eval(IO(queues -= queueName)).map(_ => ())
   override def bindQueue(channel: Channel, queueName: model.QueueName, exchangeName: model.ExchangeName, routingKey: model.RoutingKey): Stream[IO, Unit] = Stream.eval(IO.unit)
   override def bindQueue(channel: Channel, queueName: model.QueueName, exchangeName: model.ExchangeName, routingKey: model.RoutingKey, args: model.QueueBindingArgs): Stream[IO, Unit] = Stream.eval(IO.unit)
   override def bindQueueNoWait(channel: Channel, queueName: model.QueueName, exchangeName: model.ExchangeName, routingKey: model.RoutingKey, args: model.QueueBindingArgs): Stream[IO, Unit] = Stream.eval(IO.unit)
   override def unbindQueue(channel: Channel, queueName: model.QueueName, exchangeName: model.ExchangeName, routingKey: model.RoutingKey): Stream[IO, Unit] = Stream.eval(IO.unit)
   override def bindExchange(channel: Channel, destination: model.ExchangeName, source: model.ExchangeName, routingKey: model.RoutingKey, args: model.ExchangeBindingArgs): Stream[IO, Unit] = Stream.eval(IO.unit)
   override def declareExchange(channel: Channel, exchangeName: model.ExchangeName, exchangeType: ExchangeType): Stream[IO, Unit] = Stream.eval(IO.unit)
-  override def declareQueue(channel: Channel, queueName: model.QueueName): Stream[IO, Unit] = Stream.eval(IO.unit)
+  override def declareQueue(channel: Channel, queueName: model.QueueName): Stream[IO, Unit] = Stream.eval(IO(queues += queueName)).map(_ => ())
 }
