+++ examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/IOAckerConsumer.scala
@@ -18,7 +18,6 @@ package com.github.gvolpe.fs2rabbit.examples
 
 import cats.effect.IO
 import com.github.gvolpe.fs2rabbit.StreamLoop
-import com.github.gvolpe.fs2rabbit.config.Fs2RabbitConfigManager
 import com.github.gvolpe.fs2rabbit.instances.streameval._
 import com.github.gvolpe.fs2rabbit.interpreter.Fs2Rabbit
 
@@ -26,10 +25,8 @@ import scala.concurrent.ExecutionContext
 
 object IOAckerConsumer extends IOApp {
 
-  private val config = new Fs2RabbitConfigManager[IO].config
-
   implicit val appS: ExecutionContext     = scala.concurrent.ExecutionContext.Implicits.global
-  implicit val interpreter: Fs2Rabbit[IO] = new Fs2Rabbit[IO](config)
+  implicit val interpreter: Fs2Rabbit[IO] = Fs2Rabbit[IO]
 
   override def start(args: List[String]): IO[Unit] =
     StreamLoop.run(() => new AckerConsumerDemo[IO]().program)
