+++ examples/src/main/scala/com/github/gvolpe/fs2rabbit/examples/MonixAutoAckConsumer.scala
@@ -18,7 +18,6 @@ package com.github.gvolpe.fs2rabbit.examples
 
 import cats.effect.IO
 import com.github.gvolpe.fs2rabbit.StreamLoop
-import com.github.gvolpe.fs2rabbit.config.Fs2RabbitConfigManager
 import com.github.gvolpe.fs2rabbit.instances.streameval._
 import com.github.gvolpe.fs2rabbit.interpreter.Fs2Rabbit
 import monix.eval.Task
@@ -28,10 +27,8 @@ import scala.concurrent.ExecutionContext
 
 object MonixAutoAckConsumer extends IOApp {
 
-  private val config = new Fs2RabbitConfigManager[Task].config
-
   implicit val appS: ExecutionContext       = scala.concurrent.ExecutionContext.Implicits.global
-  implicit val interpreter: Fs2Rabbit[Task] = new Fs2Rabbit[Task](config)
+  implicit val interpreter: Fs2Rabbit[Task] = Fs2Rabbit[Task]
 
   override def start(args: List[String]): IO[Unit] =
     StreamLoop.run(() => new AutoAckConsumerDemo[Task].program)
